broke admin page into components
